# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: CoursesList API
  version: "1.0.0"
  summary: API systemu obsługi szkoleń — kursy, kursanci, firmy, zaświadczenia, dzienniki szkoleń.
  description: |
    Specyfikacja dla aplikacji integrujących się z API serwer-serwer za pomocą klucza API.

    **Przeczytaj najpierw `INTEGRATION.md`** z tego samego katalogu — opisuje reguły, których
    nie da się wyrazić w samym schemacie (semantyka PATCH jako pełnego nadpisania, formaty dat
    bez strefy czasowej, niespójności nazw pól, pułapki przy tworzeniu zaświadczeń).

    Konwencje:
    - Każda udana odpowiedź JSON ma kopertę `{"data": ...}`; błędy mają `{"error": {"code", "message"}}`.
    - Każda operacja wymaga zakresu podanego w `x-required-scope` (brak = wystarczy ważny klucz).
    - Operacje z `x-requires-admin: true` wymagają dodatkowo, by konto serwisowe klucza miało rolę
      administratora (`role = 1`) — sam zakres nie wystarczy.
    - Ciała żądań są odrzucane (400), jeśli zawierają pole spoza schematu.

    Wygenerowano z kodu źródłowego (`api/internal/*`) na gałęzi `api_and_webhook`.
servers:
  - url: "{baseUrl}/api/v1"
    variables:
      baseUrl:
        default: http://localhost:8081
        description: >-
          Adres Go API. Lokalnie `http://localhost:8081`; w `docker-compose.prod.yml` API nasłuchuje na
          porcie 8082. Integracje powinny wołać Go API bezpośrednio, nie przez proxy Nuxt.
security:
  - apiKey: []
...
paths:
...
  /students:
    get:
      tags: [students]
      operationId: listStudents
      summary: Lista kursantów
      description: >-
        Sortowanie: nazwisko, imię. `search` dzieli frazę na słowa (po spacjach i przecinkach);
        kursant pasuje, gdy **każde** słowo występuje w imieniu, nazwisku lub PESEL-u
        (bez rozróżniania wielkości liter, dopasowanie fragmentu).
      x-required-scope: students:read
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Limit"
        - name: companyId
          in: query
          description: Tylko kursanci przypisani do tej firmy.
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Kursanci.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [students]
      operationId: createStudent
      summary: Utworzenie kursanta
      x-required-scope: students:write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentWrite"
      responses:
        "201":
          description: Utworzony kursant.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentDetailsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: "Firma wskazana w `companyId` nie istnieje (`company not found`)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: >-
        Klucz API w nagłówku `Authorization: Bearer clk_...`. Klucze wystawia administrator w aplikacji
        webowej (Administracja → Klucze API). Klucz działa w imieniu przypisanego konta serwisowego i
        tylko w zakresie nadanych uprawnień (`x-required-scope`).
  parameters:
...
    Search:
      name: search
      in: query
      description: Fraza wyszukiwania; białe znaki na brzegach są obcinane, pusta = brak filtra.
      schema:
        type: string
    Limit:
      name: limit
      in: query
      description: Maksymalna liczba wyników. Poza zakresem 1–100 → 400 `incorrect limit value`.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...
  responses:
    BadRequest:
      description: >-
        Błędne żądanie — np. `invalid request body` (brak wymaganego pola, pole spoza schematu,
        zła data), błędny identyfikator w ścieżce, `incorrect limit value`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Unauthorized:
      description: >-
        Brak lub nieważny klucz: `missing credentials`, `invalid or expired api key`.
        Klucz mógł zostać unieważniony, wygasnąć albo jego konto usunięto.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: >-
        Klucz jest ważny, ale nie ma uprawnień: `api key is missing the <zakres> scope`
        albo `admin access required` (konto serwisowe nie jest administratorem).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalError:
      description: Błąd serwera (`internal_error`). Treść komunikatu nie jest stabilna.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Date:
      type: string
      format: date
      pattern: "^\\d{4}-\\d{2}-\\d{2}$"
      examples: ["2026-09-14"]
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [bad_request, unauthorized, invalid_credentials, forbidden, not_found, conflict, internal_error]
            message:
              type: string
              description: >-
                Komunikat po angielsku. Komunikaty wymienione w opisach operacji są stabilne i można na nich
                opierać logikę; pozostałe traktuj jako informacyjne.
...
    CompanyRef:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    StudentListItem:
      type: object
      required: [id, firstName, lastName, pesel, birthDate, company]
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        pesel:
          type: [string, "null"]
        birthDate:
          $ref: "#/components/schemas/Date"
        company:
          oneOf:
            - $ref: "#/components/schemas/CompanyRef"
            - type: "null"
    StudentListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/StudentListItem"
    StudentDetails:
      type: object
      required: [id, firstName, lastName, secondName, birthDate, birthPlace, pesel, addressStreet, addressCity, addressZip, telephone, company]
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        secondName:
          type: [string, "null"]
        birthDate:
          $ref: "#/components/schemas/Date"
        birthPlace:
          type: string
        pesel:
          type: [string, "null"]
        addressStreet:
          type: [string, "null"]
        addressCity:
          type: [string, "null"]
        addressZip:
          type: [string, "null"]
        telephone:
          type: [string, "null"]
        company:
          oneOf:
            - $ref: "#/components/schemas/CompanyRef"
            - type: "null"
    StudentDetailsResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/StudentDetails"
    StudentWrite:
      type: object
      additionalProperties: false
      description: >-
        Wymagane pola nie mogą być puste po obcięciu białych znaków. Puste stringi w polach opcjonalnych
        są zapisywane jako `null`. PESEL nie jest walidowany.
      required: [firstName, lastName, birthDate, birthPlace]
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        secondName:
          type: [string, "null"]
        birthDate:
          $ref: "#/components/schemas/Date"
        birthPlace:
          type: string
          minLength: 1
        pesel:
          type: [string, "null"]
        addressStreet:
          type: [string, "null"]
        addressCity:
          type: [string, "null"]
        addressZip:
          type: [string, "null"]
        telephone:
          type: [string, "null"]
        companyId:
          type: [integer, "null"]
          format: int64
          minimum: 1
          description: "`null` = kursant bez firmy."
...